refactor: Rename `__fixtures__` directory to `fixtures`
