+++ src/content/sync/html-to-notion/__tests__/html-to-notion.spec.ts
@@ -1,6 +1,6 @@
 import { convertHtmlToBlocks } from '../html-to-notion';
 
-import { htmlTestCases } from './__fixtures__';
+import { htmlTestCases } from './fixtures';
 
 describe('convertHtmlToBlocks', () => {
   it('throws error if HTML content does not match expected format', () => {
